requirements.txt
SQLAlchemy
Alembic
psycopg2-binary
# Environment variable management
python-dotenv
loguru
requests